veridata/veridata_rag/pyproject.toml
[project]
name = "verirag-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "pgvector>=0.2.0",
    "fastapi>=0.124.0",
    "google-generativeai>=0.8.3",
    "jinja2>=3.1.6",
    "llama-index-core>=0.14.10",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-llms-openai>=0.6.10",
    "llama-index-multi-modal-llms-gemini>=0.6.1",
    "psycopg[binary,pool]>=3.3.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"
